# Optimize working with filters and extensions in FileDialog, decrease allocations