--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/Microsoft/Win32/CommonDialog.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/Microsoft/Win32/CommonDialog.cs
@@ -28,9 +28,6 @@ namespace Microsoft.Win32
     /// <summary>
     ///  An abstract base class for displaying common dialogs.
     /// </summary>
-    /// <Remarks>
-    ///     InheritanceDemand for UIPermission (UIPermissionWindow.AllWindows)
-    /// </Remarks>
     public abstract class CommonDialog
     {
         //---------------------------------------------------
@@ -60,10 +57,7 @@ namespace Microsoft.Win32
         ///  performs initialization tasks for all common dialogs and then
         ///  calls RunDialog.
         /// </summary>
-        /// <Remarks>
-        ///     Callers must have UIPermission(UIPermissionWindow.AllWindows) to call this API.
-        /// </Remarks>
-        public virtual Nullable<bool> ShowDialog()
+        public virtual bool? ShowDialog()
         {
             CheckPermissionsToShowDialog();
 
@@ -126,10 +120,7 @@ namespace Microsoft.Win32
         /// <summary>
         ///  Runs a common dialog box, with the owner as the given Window
         /// </summary>
-        /// <Remarks>
-        ///     Callers must have UIPermission(UIPermissionWindow.AllWindows) to call this API.
-        /// </Remarks>
-        public Nullable<bool> ShowDialog(Window owner)
+        public bool? ShowDialog(Window owner)
         {
             CheckPermissionsToShowDialog();
 
@@ -373,7 +364,7 @@ namespace Microsoft.Win32
         // Private variable used to store data for the Tag property
         private object _userData;
 
-        private Thread _thread = Thread.CurrentThread;
+        private readonly Thread _thread = Thread.CurrentThread;
 
         /// <summary>
         ///  The owner hwnd passed into the dialog is stored as a private