--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/Microsoft/Win32/CommonItemDialog.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/Microsoft/Win32/CommonItemDialog.cs
@@ -697,10 +697,10 @@ namespace Microsoft.Win32
         {
             public delegate bool OnOkCallback(IFileDialog dialog);
 
-            private IFileDialog _dialog;
+            private readonly IFileDialog _dialog;
 
-            private OnOkCallback _okCallback;
-            uint _eventCookie;
+            private readonly OnOkCallback _okCallback;
+            private readonly uint _eventCookie;
 
             public VistaDialogEvents(IFileDialog dialog, OnOkCallback okCallback)
             {