--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/Microsoft/Win32/OpenFileDialog.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/Microsoft/Win32/OpenFileDialog.cs
@@ -61,9 +61,6 @@ namespace Microsoft.Win32
         /// <exception cref="System.InvalidOperationException">
         /// Thrown if there are no filenames stored in the OpenFileDialog.
         /// </exception>
-        /// <Remarks>
-        ///     Callers must have FileDialogPermission(FileDialogPermissionAccess.Open) to call this API.
-        /// </Remarks>
         public Stream OpenFile()
         {
             string filename = CriticalItemName;
@@ -85,9 +82,6 @@ namespace Microsoft.Win32
         /// <exception cref="System.InvalidOperationException">
         /// Thrown if there are no filenames stored in the OpenFileDialog
         /// </exception>
-        /// <Remarks>
-        ///     Callers must have FileDialogPermission(FileDialogPermissionAccess.Open) to call this API.
-        /// </Remarks>
         public Stream[] OpenFiles()
         {
             // Cache ItemNames to avoid perf issues as per
@@ -124,9 +118,6 @@ namespace Microsoft.Win32
         /// <summary>
         ///  Resets all properties to their default values.
         /// </summary>
-        /// <Remarks>
-        ///     Callers must have FileIOPermission(PermissionState.Unrestricted) to call this API.
-        /// </Remarks>
         public override void Reset()
         {