--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/Microsoft/Win32/OpenFileDialog.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/Microsoft/Win32/OpenFileDialog.cs
@@ -1,4 +1,4 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
@@ -197,22 +197,6 @@ namespace Microsoft.Win32
 
         //---------------------------------------------------
         //
-        // Public Events
-        //
-        //---------------------------------------------------
-        //#region Public Events
-        //#endregion Public Events
-
-        //---------------------------------------------------
-        //
-        // Protected Methods
-        //
-        //---------------------------------------------------
-        // #region Protected Methods
-        // #endregion Protected Methods
-
-        //---------------------------------------------------
-        //
         // Internal Methods
         //
         //---------------------------------------------------
@@ -224,22 +208,6 @@ namespace Microsoft.Win32
         }
 
         #endregion Internal Methods
-
-        //---------------------------------------------------
-        //
-        // Internal Properties
-        //
-        //---------------------------------------------------
-        //#region Internal Properties
-        //#endregion Internal Properties
-
-        //---------------------------------------------------
-        //
-        // Internal Events
-        //
-        //---------------------------------------------------
-        //#region Internal Events
-        //#endregion Internal Events
 
         //---------------------------------------------------
         //
@@ -268,21 +236,5 @@ namespace Microsoft.Win32
         }
 
         #endregion Private Methods
-
-        //---------------------------------------------------
-        //
-        // Private Properties
-        //
-        //---------------------------------------------------
-        //#region Private Properties
-        //#endregion Private Properties
-
-        //---------------------------------------------------
-        //
-        // Private Fields
-        //
-        //---------------------------------------------------
-        //#region Private Fields
-        //#endregion Private Fields        
     }
 }