--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/Microsoft/Win32/SaveFileDialog.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/Microsoft/Win32/SaveFileDialog.cs
@@ -60,9 +60,6 @@ namespace Microsoft.Win32
         /// <exception cref="System.InvalidOperationException">
         /// Thrown if there are no filenames stored in the SaveFileDialog.
         /// </exception>
-        /// <Remarks>
-        ///     Callers must have UIPermission.AllWindows to call this API.
-        /// </Remarks>
         public Stream OpenFile()
         {
 
@@ -87,9 +84,6 @@ namespace Microsoft.Win32
         /// <summary>
         ///  Resets all properties to their default values.
         /// </summary>
-        /// <Remarks>
-        ///     Callers must have UIPermission.AllWindows to call this API.
-        /// </Remarks>
         public override void Reset()
         {
 
@@ -124,9 +118,6 @@ namespace Microsoft.Win32
         ///  Gets or sets a value indicating whether the dialog box prompts the user for
         ///  permission to create a file if the user specifies a file that does not exist.
         /// </summary>
-        /// <Remarks>
-        ///     Callers must have UIPermission.AllWindows to call this API.
-        /// </Remarks>
         public bool CreatePrompt { get; set; }
 
         /// <summary>
@@ -156,9 +147,6 @@ namespace Microsoft.Win32
         /// Gets or sets a value indicating whether the Save As dialog box displays a 
         /// warning if the user specifies a file name that already exists.
         /// </summary>
-        /// <Remarks>
-        ///     Callers must have UIPermission.AllWindows to call this API.
-        /// </Remarks>
         public bool OverwritePrompt { get; set; }
 
         #endregion Public Properties