## Description

Decreases allocations and improves performance when working with `FileDialog` and the derived classes.
- `COMDLG_FILTERSPEC` is now a `readonly struct`, it is defined as `const` in the COM interop.
- Replaced string overloads for `StartsWith`, `IndexOf` etc. with char overload instead of single-character strings.
- `string.CompareOrdinal` replaced with `string.Equals(a, b, StringComparison.Ordinal)`
- Instead of using `string.Split` to get `Length` we count them using `MemoryExtensions.Count`
- `GetFilterItems` will no longer allocate a `List<COMDLG_FILTERSPEC>` with `ToArray()` conversion.
- `GetFilterExtensions` will no longer allocate a `List<string>` with `ToArray()` conversion.
- Removed obsolete CAS remarks on `FileDialog` and derived classes.

### Using `string.Split` vs `MemoryExtensions.Count` for counting

| Method   | findPattern          | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|--------- |--------------------- |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original | Image(...))\|\*\.\* [68] | 47.795 ns |  0.7218 ns |   0.6027 ns | 0.0172 |         672 B |         288 B |
| PR__EDIT | Image(...))\|\*\.\* [68] |  4.336 ns |  0.0174 ns |   0.0154 ns |      - |          98 B |             - |

<details><summary>Benchmark code</summary>

```csharp
[Benchmark]
[Arguments("Image Files(*.BMP;*.JPG;*.GIF)|*.BMP;*.JPG;*.GIF|All files (*.*)|*.*")]
public  int Original(string updatedFilter)
{
    string[] formats = updatedFilter.Split('|');

    return formats.Length; 
}

[Benchmark]
[Arguments("Image Files(*.BMP;*.JPG;*.GIF)|*.BMP;*.JPG;*.GIF|All files (*.*)|*.*")]
public int PR__EDIT(string updatedFilter)
{
    return findPattern.AsSpan().Count("|");
}
```

</details>

### GetFilterItems benchmark

| Method   | filter               | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|--------- |--------------------- |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original | Image(...))\|\*.\* [68] |  80.42 ns |   1.255 ns |    1.113 ns | 0.0277 |       2,413 B |         464 B |
| PR__EDIT | Image(...))\|\*.\* [68] |  58.19 ns |   0.809 ns |    0.757 ns | 0.0206 |         870 B |         344 B |
| Original | Imag(...)\|\*.\* [206] |  213.0 ns |    4.28 ns |     5.71 ns | 0.0720 |       2,402 B |        1208 B |
| PR__EDIT | Imag(...)\|\*.\* [206] |  158.0 ns |    3.19 ns |     4.97 ns | 0.0558 |         870 B |         936 B |

<details><summary>Benchmark code</summary>

```csharp
[Benchmark]
[Arguments("Image Files(*.BMP;*.JPG;*.GIF)|*.BMP;*.JPG;*.GIF|All files (*.*)|*.*")]
public COMDLG_FILTERSPEC[] Original(string filter)
{
    var extensions = new List<COMDLG_FILTERSPEC>();

    if (!string.IsNullOrEmpty(filter))
    {
        string[] tokens = filter.Split('|');
        if (0 == tokens.Length % 2)
        {
            for (int i = 1; i < tokens.Length; i += 2)
            {
                 extensions.Add(new COMDLG_FILTERSPEC_OLD() { pszSpec = tokens[i - 1], pszName = tokens[i] });
            }
        }
    }
    return extensions.ToArray();
}

  [Benchmark]
  [Arguments("Image Files(*.BMP;*.JPG;*.GIF)|*.BMP;*.JPG;*.GIF|All files (*.*)|*.*")]
  public COMDLG_FILTERSPEC[] PR__EDIT(string filter)
  {
      COMDLG_FILTERSPEC[] extensions = null;

      if (!string.IsNullOrEmpty(filter))
      {
          string[] tokens = filter.Split('|');
          if (tokens.Length % 2 == 0)
          {
              extensions = new COMDLG_FILTERSPEC[tokens.Length / 2];
              for (int i = 0; i < extensions.Length; i++)
              {
                  extensions[i] = new(tokens[i * 2], tokens[i * 2 + 1]);
              }
          }
      }

      return extensions ?? Array.Empty<COMDLG_FILTERSPEC>();
  }
```

</details>

## Customer Impact

Improved performance, decreased allocations.

## Regression

No.

## Testing

Local build, testing with file dialogs.

## Risk

Low.

 ###### Microsoft Reviewers: [Open in CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/dotnet/wpf/pull/9599)